// models/pixel.ts
export interface Pixel {
  elevation: number;
  df_ele_cnty_name: string;
  bmcwn_0: number;
  bmcwn_15: number;
  bmcwn_2: number;
  bmcwn_25: number;
  bmcwn_35: number;
  bmcwn_40: number;
  bmcwn_7: number;
  bmfol_0: number;
  bmfol_15: number;
  bmfol_2: number;
  bmfol_25: number;
  bmfol_35: number;
  bmfol_40: number;
  bmfol_7: number;
  bmstm_0: number;
  bmstm_15: number;
  bmstm_2: number;
  bmstm_25: number;
  bmstm_35: number;
  bmstm_40: number;
  bmstm_7: number;
  sng_0: number;
  sng_15: number;
  sng_2: number;
  sng_25: number;
  sng_35: number;
  sng_40: number;
  sng_7: number;
  tpa_0: number;
  tpa_15: number;
  tpa_2: number;
  tpa_25: number;
  tpa_35: number;
  tpa_40: number;
  tpa_7: number;
  cluster1: number;
  cluster2: number;
  x: number;
  y: number;
  cluster_no: number;
}
// tslint:disable: variable-name
export class Pixel implements Pixel {
  elevation = 0;
  df_ele_cnty_name = '';
  bmcwn_0 = 0;
  bmcwn_15 = 0;
  bmcwn_2 = 0;
  bmcwn_25 = 0;
  bmcwn_35 = 0;
  bmcwn_40 = 0;
  bmcwn_7 = 0;
  bmfol_0 = 0;
  bmfol_15 = 0;
  bmfol_2 = 0;
  bmfol_25 = 0;
  bmfol_35 = 0;
  bmfol_40 = 0;
  bmfol_7 = 0;
  bmstm_0 = 0;
  bmstm_15 = 0;
  bmstm_2 = 0;
  bmstm_25 = 0;
  bmstm_35 = 0;
  bmstm_40 = 0;
  bmstm_7 = 0;
  sng_0 = 0;
  sng_15 = 0;
  sng_2 = 0;
  sng_25 = 0;
  sng_35 = 0;
  sng_40 = 0;
  sng_7 = 0;
  tpa_0 = 0;
  tpa_15 = 0;
  tpa_2 = 0;
  tpa_25 = 0;
  tpa_35 = 0;
  tpa_40 = 0;
  tpa_7 = 0;
  cluster1 = 0;
  cluster2 = 0;
  x = 0;
  y = 0;
  cluster_no = 0;
}